requirements.txt
flask>=3.0
PyJWT>=2.8
cryptography>=42.0
